HTML5 input type=numeric "change" fix
The editor correctly fires the "change" event if the user types
characters in the field with the keyboard but if the user uses the
up/down arrows to change the numeric value the "change" event would not
fire. I know doubling up onKeyPress isn't ideal but it solves the
problem.