Allow schema title to be a function
It is usefull when you want to compute data in the label. For instance I need to do: Price ($) where the currency is variable. And the variable may not be available when the model is instantiated, though it will when the form is rendered. So I cannot just do `title: 'text'+variable`